+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 ### Added
 
-- [Create Attendace report](https://gitlab.com/atri-tech/atri-maintainers/ethal/-/merge_requests/218)
-- [Create multiselect dialog ui for updating asset maintenance log status](https://gitlab.com/atri-tech/atri-maintainers/ethal/-/merge_requests/217)
+- [Create Attendace report | Create multiselect dialog ui for updating asset maintenance log status](https://gitlab.com/atri-tech/atri-maintainers/ethal/-/merge_requests/225)
 
 ## [v1.3.1] - 2021-09-25