+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- [Create Attendace report | Create multiselect dialog ui for updating asset maintenance log status](https://gitlab.com/atri-tech/atri-maintainers/ethal/-/merge_requests/225)
+
+## [v1.3.1] - 2021-09-25
+
 ### Fixes
 
 - [fix purchase order submission issue](https://gitlab.com/atri-tech/atri-maintainers/ethal/-/merge_requests/214)